# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/packages": {
      "get": {
        "operationId": "PackagesPublicController_list",
        "summary": "Danh sách gói đang bán (isActive=true)",
        "description": "Query `type`: INTERNET | SPEEDX | FPT_PLAY | CAMERA | SERVICE. Response: mảng PackagePublicDto.",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "INTERNET",
                "SPEEDX",
                "FPT_PLAY",
                "CAMERA",
                "SERVICE"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PackagePublicDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Public — Packages"
        ]
      }
    },
    "/api/v1/packages/by-code/{code}": {
      "get": {
        "operationId": "PackagesPublicController_byCode",
        "summary": "Chi tiết gói theo code (vd. internet-giga)",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "example": "internet-giga",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackagePublicDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Public — Packages"
        ]
      }
    },
    "/api/v1/packages/{id}": {
      "get": {
        "operationId": "PackagesPublicController_byId",
        "summary": "Chi tiết gói theo MongoDB id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackagePublicDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Public — Packages"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "MeController_getMe",
        "summary": "Profile khách đang đăng nhập",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me (Khách)"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "MeController_patchMe",
        "summary": "Cập nhật profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me (Khách)"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/me/leads": {
      "get": {
        "operationId": "MeController_myLeads",
        "summary": "Đơn đã gửi khi đăng nhập (customerId)",
        "description": "Khác GET /leads/history — chỉ đơn gắn tài khoản.",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyLeadsListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me (Khách)"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/api/v1/package-quiz": {
      "get": {
        "operationId": "PackageQuizPublicController_getActive",
        "summary": "Lấy bộ câu hỏi gợi ý (wizard 2–3 câu)",
        "description": "Query `code=home-needs`. Trả questions[], mỗi câu có multiSelect và options (label, icon, code).",
        "parameters": [
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Mã bộ câu hỏi — mặc định quiz visible đầu tiên",
            "schema": {
              "example": "home-needs",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageQuizPublicDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Public — Package quiz"
        ]
      }
    },
    "/api/v1/package-quiz/recommend": {
      "post": {
        "operationId": "PackageQuizPublicController_recommend",
        "summary": "Nộp đủ câu trả lời → gợi ý loại gói + danh sách gói",
        "description": "Body `answers[]` (2–3 câu). Response: recommendedTypes, packages[], resultsPath cho FE navigate.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendPackageQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageQuizRecommendResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Public — Package quiz"
        ]
      }
    },
...
  },
  "info": {
    "title": "FPT Telecom Landing API",
    "description": "REST API cho landing page & admin FPT Telecom.\n\n**Base path:** `/api/v1`\n\n**Xác thực:**\n- Access token: header `Authorization: Bearer <accessToken>`\n- Refresh token: cookie HttpOnly `refreshToken` (credentials: include)\n\n**FE public (không bắt buộc login):** packages, leads, faqs, navigation, package-quiz, banners\n\n**Rate limit:** leads history, lead create theo IP/SĐT\n\n**Swagger UI:** `/api/docs` · **OpenAPI JSON:** `/api/docs-json`",
    "version": "1.0.0",
    "contact": {
      "name": "FPT Telecom BE",
      "url": "",
      "email": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "accessToken từ login/register/refresh"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "Refresh JWT — tự gửi khi credentials: true"
      }
    },
    "schemas": {
...
      "AuthUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "defaultAddress": {
            "type": "object",
            "nullable": true
          },
          "role": {
            "type": "string",
            "example": "CUSTOMER"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "fullName",
          "role"
        ]
      },
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Số điện thoại không hợp lệ"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "PackagePublicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "internet-giga"
          },
          "name": {
            "type": "string"
          },
          "tagline": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "INTERNET",
              "SPEEDX",
              "FPT_PLAY",
              "CAMERA",
              "SERVICE"
            ]
          },
          "price": {
            "type": "number"
          },
          "heroImage": {
            "type": "string"
          },
          "bannerImage": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "description": "Object metadata (downloadMbps, audience, …)"
          }
        },
        "required": [
          "id",
          "code",
          "name"
        ]
      },
...
      "PackageSnapshotDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "object",
            "nullable": true
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "type"
        ]
      },
...
      "LeadPublicItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "installAddress": {
            "type": "string"
          },
          "packageId": {
            "type": "object",
            "nullable": true
          },
          "packageSnapshot": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/PackageSnapshotDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "createdAt",
          "fullName",
          "phone",
          "installAddress"
        ]
      },
...
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "defaultAddress": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true,
            "description": "Gửi null hoặc chuỗi rỗng để xóa email"
          }
        }
      },
      "MyLeadsListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeadPublicItemDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
...
      "PackageQuizOptionPublicDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "displayOrder": {
            "type": "number"
          }
        },
        "required": [
          "code",
          "label",
          "icon",
          "displayOrder"
        ]
      },
      "PackageQuizQuestionPublicDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "multiSelect": {
            "type": "boolean"
          },
          "displayOrder": {
            "type": "number"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageQuizOptionPublicDto"
            }
          }
        },
        "required": [
          "code",
          "title",
          "multiSelect",
          "displayOrder",
          "options"
        ]
      },
      "PackageQuizPublicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "home-needs"
          },
          "tagline": {
            "type": "object",
            "nullable": true
          },
          "icon": {
            "type": "string"
          },
          "questionCount": {
            "type": "number",
            "example": 3
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageQuizQuestionPublicDto"
            }
          }
        },
        "required": [
          "id",
          "code",
          "icon",
          "questionCount",
          "questions"
        ]
      },
      "QuizAnswerDto": {
        "type": "object",
        "properties": {
          "questionCode": {
            "type": "string",
            "example": "wifi-usage"
          },
          "optionCodes": {
            "example": [
              "media",
              "game"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "questionCode",
          "optionCodes"
        ]
      },
      "RecommendPackageQuizDto": {
        "type": "object",
        "properties": {
          "quizCode": {
            "type": "string",
            "example": "home-needs"
          },
          "answers": {
            "description": "Trả lời từng câu hỏi — ưu tiên dùng field này",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizAnswerDto"
            }
          },
          "questionCode": {
            "type": "string",
            "example": "wifi-usage"
          },
          "optionCodes": {
            "example": [
              "study"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PackageQuizAnswerSummaryDto": {
        "type": "object",
        "properties": {
          "questionCode": {
            "type": "string"
          },
          "optionCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "questionCode",
          "optionCodes"
        ]
      },
      "PackageQuizRankingDto": {
        "type": "object",
        "properties": {
          "packageType": {
            "type": "string",
            "enum": [
              "INTERNET",
              "SPEEDX",
              "FPT_PLAY",
              "CAMERA",
              "SERVICE"
            ]
          },
          "score": {
            "type": "number"
          },
          "label": {
            "type": "string"
          },
          "sectionId": {
            "type": "string",
            "example": "#internet"
          }
        },
        "required": [
          "packageType",
          "score",
          "label",
          "sectionId"
        ]
      },
      "PackageQuizRecommendResponseDto": {
        "type": "object",
        "properties": {
          "quizCode": {
            "type": "string",
            "example": "home-needs"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageQuizAnswerSummaryDto"
            }
          },
          "recommendedTypes": {
            "example": [
              "FPT_PLAY",
              "INTERNET"
            ],
            "description": "INTERNET | SPEEDX | FPT_PLAY | CAMERA | SERVICE",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "primaryType": {
            "type": "string",
            "enum": [
              "INTERNET",
              "SPEEDX",
              "FPT_PLAY",
              "CAMERA",
              "SERVICE"
            ]
          },
          "scrollToSection": {
            "type": "string",
            "example": "#truyen-hinh"
          },
          "resultsPath": {
            "type": "string",
            "example": "/ket-qua-tu-van?types=FPT_PLAY,INTERNET&primary=FPT_PLAY",
            "description": "FE navigate tới màn hình danh sách gói gợi ý"
          },
          "message": {
            "type": "string"
          },
          "rankings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageQuizRankingDto"
            }
          },
          "packages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackagePublicDto"
            }
          },
          "totalPackages": {
            "type": "number"
          }
        },
        "required": [
          "quizCode",
          "answers",
          "recommendedTypes",
          "primaryType",
          "scrollToSection",
          "resultsPath",
          "message",
          "rankings",
          "packages",
          "totalPackages"
        ]
      },
...
    }
  }
}